--- .idea/runConfigurations/All_Tests.xml
@@ -1,14 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="All Tests" type="mocha-javascript-test-runner">
-    <node-interpreter>project</node-interpreter>
-    <node-options />
-    <mocha-package>$PROJECT_DIR$/node_modules/mocha</mocha-package>
-    <working-directory>$PROJECT_DIR$/packages/frontend</working-directory>
-    <pass-parent-env>true</pass-parent-env>
-    <ui>bdd</ui>
-    <extra-mocha-options>--require ts-node/register</extra-mocha-options>
-    <test-kind>PATTERN</test-kind>
-    <test-pattern>src/scripts/test/**/*test.ts src/scripts/test/**/*tests.ts</test-pattern>
-    <method v="2" />
-  </configuration>
-</component>
