+++ package.json
@@ -11,7 +11,8 @@
     "packages/common-ui",
     "packages/frontend",
     "packages/math-frontend",
-    "packages/backend-resolver"
+    "packages/backend-resolver",
+    "packages/data-api-client"
   ],
   "devDependencies": {
     "@eslint/js": "^9.3.0",
@@ -19,6 +20,7 @@
     "eslint": "^8.57.0",
     "http-server": "^14.1.1",
     "nyc": "^15.1.0",
+    "swagger-typescript-api": "^13.0.21",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.5",
     "typescript-eslint": "^7.10.0",
@@ -29,6 +31,7 @@
     "docs": "npx --workspaces typedoc --json build/docs.json --exclude '**/scripts/test/**/*.ts' --excludePrivate --entryPointStrategy expand src/ && npx typedoc --entryPointStrategy merge packages/*/build/docs.json --readme API_DOC.md --out packages/frontend/dist/docs",
     "test": "npm run --workspaces --if-present test",
     "serve": "npx -w @xivgear/gearplan-frontend webpack serve",
-    "lint": "npx -ws eslint ./src/**/*.ts --exit-on-fatal-error --max-warnings 0"
+    "lint": "npx -ws eslint ./src/**/*.ts --exit-on-fatal-error --max-warnings 0",
+    "dataschema": "npx -w @xivgear/data-api-client swagger-typescript-api -o ./src/ -n dataapi.ts -p http://localhost:8085/swagger/xivgear-reference-data-api-0.1.yml --api-class-name DataApiClient"
   }
 }
