+++ packages/core/src/gear.ts
@@ -47,18 +47,18 @@ import {
     XivCombatItem
 } from "@xivgear/xivmath/geartypes";
 import {xivApiIconUrl} from "./external/xivapi";
+import {BaseParamToStatKey, RelevantBaseParam} from "./external/xivapitypes";
+import {Inactivitytimer} from "./util/inactivitytimer";
+import {addStats, finalizeStats} from "@xivgear/xivmath/xivstats";
+import {GearPlanSheet} from "./sheet";
+import {requireArrayTyped, requireBool, requireNumber, requireString} from "./external/data_validators";
 import {
     IlvlSyncInfo,
     XivApiFoodDataRaw,
     XivApiFoodItemDataRaw,
     XivApiItemDataRaw,
     XivApiMateriaDataRaw
-} from "./datamanager";
-import {BaseParamToStatKey, RelevantBaseParam} from "./external/xivapitypes";
-import {Inactivitytimer} from "./util/inactivitytimer";
-import {addStats, finalizeStats} from "@xivgear/xivmath/xivstats";
-import {GearPlanSheet} from "./sheet";
-import {requireArrayTyped, requireBool, requireNumber, requireString} from "./external/data_validators";
+} from "./datamanager_xivapi";
 
 
 export function nonEmptyRelicStats(stats: RelicStats | undefined): boolean {
@@ -701,408 +701,4 @@ export type ItemSingleStatDetail = {
     cap: 0,
 };
 
-// noinspection RedundantIfStatementJS
-export class XivApiGearInfo implements GearItem {
-    id: number;
-    name: string;
-    iconUrl: URL;
-    ilvl: number;
-    displayGearSlot: DisplayGearSlot;
-    displayGearSlotName: DisplayGearSlotKey;
-    occGearSlotName: OccGearSlotKey;
-    stats: RawStats;
-    primarySubstat: keyof RawStats | null;
-    secondarySubstat: keyof RawStats | null;
-    materiaSlots: MateriaSlot[];
-    isCustomRelic: boolean;
-    isUnique: boolean;
-    unsyncedVersion: XivApiGearInfo;
-    acquisitionType: GearAcquisitionSource;
-    statCaps: {
-        [K in RawStatKey]?: number
-    };
-    isSyncedDown: boolean;
-    relicStatModel: RelicStatModel;
-
-    constructor(data: XivApiItemDataRaw) {
-        this.id = requireNumber(data.ID);
-        this.name = requireString(data.Name);
-        this.ilvl = requireNumber(data.LevelItem['value']);
-        this.iconUrl = new URL(xivApiIconUrl(requireNumber(data.Icon['id']), true));
-        const eqs = data.EquipSlotCategory['fields'];
-        if (!eqs) {
-            console.error('EquipSlotCategory was null!', data);
-        }
-        else if (eqs['MainHand']) {
-            this.displayGearSlotName = 'Weapon';
-            if (eqs['OffHand']) {
-                this.occGearSlotName = 'Weapon2H'
-            }
-            else {
-                this.occGearSlotName = 'Weapon1H';
-            }
-        }
-        else if (eqs['OffHand']) {
-            this.displayGearSlotName = 'OffHand';
-            this.occGearSlotName = 'OffHand';
-        }
-        else if (eqs['Head']) {
-            this.displayGearSlotName = 'Head';
-            this.occGearSlotName = 'Head';
-        }
-        else if (eqs['Body']) {
-            this.displayGearSlotName = 'Body';
-            this.occGearSlotName = 'Body';
-        }
-        else if (eqs['Gloves']) {
-            this.displayGearSlotName = 'Hand';
-            this.occGearSlotName = 'Hand';
-        }
-        else if (eqs['Legs']) {
-            this.displayGearSlotName = 'Legs';
-            this.occGearSlotName = 'Legs';
-        }
-        else if (eqs['Feet']) {
-            this.displayGearSlotName = 'Feet';
-            this.occGearSlotName = 'Feet';
-        }
-        else if (eqs['Ears']) {
-            this.displayGearSlotName = 'Ears';
-            this.occGearSlotName = 'Ears';
-        }
-        else if (eqs['Neck']) {
-            this.displayGearSlotName = 'Neck';
-            this.occGearSlotName = 'Neck';
-        }
-        else if (eqs['Wrists']) {
-            this.displayGearSlotName = 'Wrist';
-            this.occGearSlotName = 'Wrist';
-        }
-        else if (eqs['FingerL'] || eqs['FingerR']) {
-            this.displayGearSlotName = 'Ring';
-            this.occGearSlotName = 'Ring';
-        }
-        else {
-            console.error("Unknown slot data!", eqs);
-        }
-        this.displayGearSlot = this.displayGearSlotName ? DisplayGearSlotInfo[this.displayGearSlotName] : undefined;
-        const weaponDelayRaw = requireNumber(data.Delayms);
-        this.stats = new RawStats();
-        this.stats.wdPhys = requireNumber(data.DamagePhys);
-        this.stats.wdMag = requireNumber(data.DamageMag);
-        this.stats.weaponDelay = weaponDelayRaw ? (weaponDelayRaw / 1000.0) : 0;
-        for (const i in requireArrayTyped(data.BaseParam, 'object')) {
-            const paramName = requireString(data.BaseParam[i]['fields']['Name']);
-            if (!paramName) {
-                continue;
-            }
-            const paramValue = requireNumber(data.BaseParamValue[i]);
-            this.stats[BaseParamToStatKey[paramName]] = paramValue;
-        }
-        if (data.CanBeHq) {
-            for (const i in requireArrayTyped(data.BaseParamSpecial, 'object')) {
-                const paramId = requireNumber(data.BaseParamSpecial[i]['value']);
-                if (paramId <= 0) {
-                    continue;
-                }
-                const paramName = requireString(data.BaseParamSpecial[i]['fields']['Name']);
-                const paramValue = requireNumber(data.BaseParamValueSpecial[i]);
-                if (paramId === 12) {
-                    this.stats.wdPhys += paramValue;
-                }
-                else if (paramId === 13) {
-                    this.stats.wdMag += paramValue;
-                }
-                else {
-                    const paramKey = BaseParamToStatKey[paramName];
-                    if (paramKey !== null) {
-                        this.stats[paramKey] += paramValue;
-                    }
-                }
-            }
-        }
-        this.isUnique = Boolean(data.IsUnique);
-        this.computeSubstats();
-        this.materiaSlots = [];
-        const baseMatCount: number = requireNumber(data.MateriaSlotCount);
-        if (baseMatCount === 0) {
-            // If there are no materia slots, then it might be a custom relic
-            // TODO: is this branch still needed?
-            if (this.displayGearSlot !== DisplayGearSlotInfo.OffHand) {
-                // Offhands never have materia slots
-                this.isCustomRelic = true;
-            }
-            else if (!this.primarySubstat) {
-                // If there is no primary substat on the item, then consider it a relic
-                this.isCustomRelic = true;
-            }
-            else {
-                // Otherwise, it's just a random item that just so happens to not have materia slots
-                this.isCustomRelic = false;
-            }
-        }
-        else {
-            // If it has materia slots, it definitely isn't a custom relic
-            this.isCustomRelic = false;
-            // Is overmelding allowed?
-            const overmeld: boolean = requireBool(data.IsAdvancedMeldingPermitted);
-            // The materia slot count represents slots that are always meldable
-            for (let i = 0; i < baseMatCount; i++) {
-                // TODO: figure out grade automatically
-                this.materiaSlots.push({
-                    maxGrade: MATERIA_LEVEL_MAX_NORMAL,
-                    allowsHighGrade: true
-                });
-            }
-            if (overmeld) {
-                // If we can also overmeld, then push a *single* high-grade slot, and then the rest are only for
-                // small materia.
-                this.materiaSlots.push({
-                    maxGrade: MATERIA_LEVEL_MAX_NORMAL,
-                    allowsHighGrade: true
-                });
-                for (let i = this.materiaSlots.length; i < MATERIA_SLOTS_MAX; i++) {
-                    this.materiaSlots.push({
-                        maxGrade: MATERIA_LEVEL_MAX_OVERMELD,
-                        allowsHighGrade: false
-                    });
-                }
-            }
-        }
-        // Try to guess the acquisition source of an item
-        // Disabled since new xivapi does not provide this in an easy form
-        // try {
-        //     const rarity: number = data.Rarity as number;
-        //     // Check if it is craftable by checking if any recipes result in this item
-        //     const isCraftable = data.GameContentLinks?.['Recipe'];
-        //     // Check if it is buyable by checking if any shops sell it
-        //     const hasShop = data.GameContentLinks?.['SpecialShop'];
-        //     // TODO: nothing results in 'normraid'
-        //     switch (rarity) {
-        //         case 1:
-        //             if (isCraftable) {
-        //                 this.acquisitionType = 'crafted';
-        //             }
-        //             break;
-        //         // Green
-        //         case 2:
-        //             if (this.name.includes('Augmented')) {
-        //                 this.acquisitionType = 'augcrafted';
-        //             }
-        //             else {
-        //                 if (isCraftable) {
-        //                     this.acquisitionType = 'crafted';
-        //                 }
-        //                 else {
-        //                     this.acquisitionType = 'dungeon';
-        //                 }
-        //             }
-        //             break;
-        //         // Blue
-        //         case 3: {
-        //             // TODO: how to differentiate raid vs tome vs aug tome?
-        //             // Aug tome: it has "augmented" in the name, easy
-        //             const isWeaponOrOH = this.occGearSlotName === 'Weapon2H' || this.occGearSlotName === 'Weapon1H' || this.occGearSlotName === 'OffHand';
-        //             if (ARTIFACT_ITEM_LEVELS.includes(this.ilvl)) {
-        //                 // Ambiguous due to start-of-expac ex trials
-        //                 if (isWeaponOrOH) {
-        //                     this.acquisitionType = 'other';
-        //                 }
-        //                 else {
-        //                     this.acquisitionType = 'artifact';
-        //                 }
-        //             }
-        //             // Start-of-expac uncapped tome gear
-        //             else if (BASIC_TOME_GEAR_ILVLS.includes(this.ilvl)) {
-        //                 this.acquisitionType = 'tome';
-        //             }
-        //             const chkRelIlvl = (relativeToRaidTier: number) => {
-        //                 return RAID_TIER_ILVLS.includes(this.ilvl - relativeToRaidTier);
-        //             };
-        //             if (chkRelIlvl(0)) {
-        //                 if (this.name.includes('Augmented')) {
-        //                     this.acquisitionType = 'augtome';
-        //                 }
-        //                 else {
-        //                     this.acquisitionType = 'raid';
-        //                 }
-        //             }
-        //             else if (chkRelIlvl(-10)) {
-        //                 if (hasShop) {
-        //                     this.acquisitionType = 'tome';
-        //                 }
-        //                 else {
-        //                     this.acquisitionType = 'alliance';
-        //                 }
-        //             }
-        //             else if (chkRelIlvl(-20)) {
-        //
-        //                 if (isWeaponOrOH) {
-        //                     this.acquisitionType = 'extrial';
-        //                 }
-        //                 else {
-        //                     // Ambiguous - first-of-the-expac extreme trial accessories and normal raid
-        //                     // accessories share the same ilvl
-        //                     this.acquisitionType = 'other';
-        //                 }
-        //             }
-        //             else if ((chkRelIlvl(-5) || chkRelIlvl(-15)) && isWeaponOrOH) {
-        //                 this.acquisitionType = 'extrial';
-        //             }
-        //             else if (this.ilvl % 10 === 5) {
-        //                 if (isWeaponOrOH) {
-        //                     if (chkRelIlvl(5)) {
-        //                         if (this.name.includes('Ultimate')) {
-        //                             this.acquisitionType = 'ultimate';
-        //                         }
-        //                         else {
-        //                             this.acquisitionType = 'raid';
-        //                         }
-        //                     }
-        //                 }
-        //             }
-        //             break;
-        //         }
-        //         // Purple
-        //         case 4:
-        //             this.acquisitionType = 'relic';
-        //             break;
-        //     }
-        // }
-        // catch (e) {
-        //     console.error("Error determining item rarity", data);
-        // }
-        // if (!this.acquisitionType) {
-        //     console.warn(`Unable to determine acquisition source for item ${this.name} (${this.id})`, data);
-        //     this.acquisitionType = 'other';
-        // }
-        this.acquisitionType = 'other';
-    }
-
-    private computeSubstats() {
-        const sortedStats = Object.entries({
-            crit: this.stats.crit,
-            dhit: this.stats.dhit,
-            determination: this.stats.determination,
-            spellspeed: this.stats.spellspeed,
-            skillspeed: this.stats.skillspeed,
-            piety: this.stats.piety,
-            tenacity: this.stats.tenacity,
-        })
-            .sort((left, right) => {
-                if (left[1] > right[1]) {
-                    return 1;
-                }
-                else if (left[1] < right[1]) {
-                    return -1;
-                }
-                return 0;
-            })
-            .filter(item => item[1])
-            .reverse();
-        if (sortedStats.length < 2) {
-            this.primarySubstat = null;
-            this.secondarySubstat = null;
-        }
-        else {
-            this.primarySubstat = sortedStats[0][0] as keyof RawStats;
-            this.secondarySubstat = sortedStats[1][0] as keyof RawStats;
-        }
-    }
-
-    /**
-     * TODO fix docs for this
-     */
-    applyIlvlData(nativeIlvlInfo: IlvlSyncInfo, syncIlvlInfo?: IlvlSyncInfo) {
-        const statCapsNative = {};
-        Object.entries(this.stats).forEach(([stat, _]) => {
-            statCapsNative[stat] = nativeIlvlInfo.substatCap(this.occGearSlotName, stat as RawStatKey);
-        });
-        this.statCaps = statCapsNative;
-        if (syncIlvlInfo && syncIlvlInfo.ilvl < this.ilvl) {
-            this.unsyncedVersion = {
-                ...this
-            };
-            this.materiaSlots = [];
-            const statCapsSync = {};
-            Object.entries(this.stats).forEach(([stat, v]) => {
-                statCapsSync[stat] = syncIlvlInfo.substatCap(this.occGearSlotName, stat as RawStatKey);
-            });
-            this.stats = applyStatCaps(this.stats, statCapsSync);
-            this.statCaps = statCapsSync;
-            this.computeSubstats();
-            this.isSyncedDown = true;
-        }
-        else {
-            this.unsyncedVersion = this;
-            this.isSyncedDown = false;
-        }
-    }
-}
-
-export class XivApiFoodInfo implements FoodItem {
-    bonuses: {
-        [K in RawStatKey]?: {
-            percentage: number;
-            max: number
-        }
-    } = {};
-    iconUrl: URL;
-    id: number;
-    name: string;
-    ilvl: number;
-    primarySubStat: RawStatKey | undefined;
-    secondarySubStat: RawStatKey | undefined;
-
-    constructor(data: XivApiFoodDataRaw, foodData: XivApiFoodItemDataRaw) {
-        this.id = requireNumber(data.ID);
-        this.name = requireString(data.Name);
-        this.iconUrl = new URL(xivApiIconUrl(requireNumber(data.Icon['id'])));
-        this.ilvl = requireNumber(data.LevelItem['value']);
-        for (const index in requireArrayTyped(foodData.BaseParam, 'object')) {
-            const baseParamName = requireString(foodData['BaseParam'][index].fields.Name) as RelevantBaseParam;
-            this.bonuses[BaseParamToStatKey[baseParamName]] = {
-                percentage: requireNumber(foodData.ValueHQ[index]),
-                max: requireNumber(foodData.MaxHQ[index]),
-            }
-        }
-        const sortedStats = Object.entries(this.bonuses).sort((entryA, entryB) => entryB[1].max - entryA[1].max).map(entry => entry[0] as RawStatKey).filter(stat => stat !== 'vitality');
-        if (sortedStats.length >= 1) {
-            this.primarySubStat = sortedStats[0];
-        }
-        if (sortedStats.length >= 2) {
-            this.secondarySubStat = sortedStats[1];
-        }
-    }
-
-}
-
-export function processRawMateriaInfo(data: XivApiMateriaDataRaw): Materia[] {
-    const out: Materia[] = [];
-    for (let i = MATERIA_LEVEL_MIN_RELEVANT - 1; i < MATERIA_LEVEL_MAX_NORMAL; i++) {
-        const itemData = data.Item[i];
-        const itemFields = itemData['fields'];
-        const itemId = requireNumber(itemData['row_id']);
-        const itemName = requireString(itemFields["Name"]);
-        const stats = new RawStats();
-        // TODO: use statById in more places
-        const stat = statById(requireNumber(data.BaseParam['row_id']));
-        if (!stat || !itemName) {
-            continue;
-        }
-        stats[stat] = requireNumber(data.Value[i]);
-        const grade = (i + 1);
-        out.push({
-            name: itemName,
-            id: itemId,
-            iconUrl: new URL(xivApiIconUrl(requireNumber(itemFields['Icon']['id']), true)),
-            stats: stats,
-            primaryStat: stat,
-            primaryStatValue: stats[stat],
-            materiaGrade: grade,
-            isHighGrade: (grade % 2) === 0
-        });
-    }
-    return out;
-}
