+++ packages/core/src/test/datamanager_tests.ts
@@ -1,8 +1,9 @@
 import 'global-jsdom/register'
 import 'isomorphic-fetch'
-import {DataManager} from "../datamanager";
 import * as assert from "assert";
 import {RawStats} from "@xivgear/xivmath/geartypes";
+import {NewApiDataManager} from "../datamanager_new";
+import {XivApiDataManager} from "../datamanager_xivapi";
 
 function eq<T>(actual: T, expected: T) {
     assert.equal(actual, expected);
@@ -12,9 +13,9 @@ function deq<T>(actual: T, expected: T) {
     assert.deepEqual(actual, expected);
 }
 
-describe('Datamanager', () => {
+describe('Old Datamanager', () => {
     it('can load some SCH items', async () => {
-        const dm = new DataManager('SCH', 90);
+        const dm = new XivApiDataManager('SCH', 90);
         await dm.loadData();
         const codexOfAscension = dm.itemById(40176);
         // Basic item props
@@ -82,7 +83,92 @@ describe('Datamanager', () => {
 
     }).timeout(20_000);
     it('can get stats of food items', async () => {
-        const dm = new DataManager('SCH', 90);
+        const dm = new XivApiDataManager('SCH', 90);
+        await dm.loadData();
+        const food = dm.foodById(44096);
+        eq(food.id, 44096);
+        eq(food.name, "Vegetable Soup");
+        eq(food.primarySubStat, 'dhit');
+        eq(food.secondarySubStat, 'determination');
+        eq(food.bonuses.dhit.max, 121);
+        eq(food.bonuses.dhit.percentage, 10);
+        eq(food.bonuses.determination.max, 73);
+        eq(food.bonuses.determination.percentage, 10);
+    }).timeout(20_000);
+});
+
+describe('New Datamanager', () => {
+    it('can load some SCH items', async () => {
+        const dm = new NewApiDataManager('SCH', 90);
+        await dm.loadData();
+        const codexOfAscension = dm.itemById(40176);
+        // Basic item props
+        eq(codexOfAscension.id, 40176);
+        eq(codexOfAscension.name, 'Codex of Ascension');
+        // TODO: fix the extra / ?
+        eq(codexOfAscension.iconUrl.toString(), 'https://beta.xivapi.com/api/1/asset/ui/icon/033000/033387_hr1.tex?format=png');
+
+        // XivCombatItem props
+        deq(codexOfAscension.stats, new RawStats({
+            wdPhys: 132,
+            wdMag: 132,
+            mind: 416,
+            crit: 306,
+            determination: 214,
+            vitality: 412,
+            weaponDelay: 3.12
+        }));
+
+        // GearItem props
+        eq(codexOfAscension.displayGearSlotName, 'Weapon');
+        eq(codexOfAscension.occGearSlotName, 'Weapon2H');
+        eq(codexOfAscension.ilvl, 665);
+        eq(codexOfAscension.primarySubstat, 'crit');
+        eq(codexOfAscension.secondarySubstat, 'determination');
+
+        deq(codexOfAscension.statCaps, {
+            // Primary stats
+            strength: 416,
+            dexterity: 416,
+            intelligence: 416,
+            mind: 416,
+
+            // Substats
+            crit: 306,
+            determination: 306,
+            dhit: 306,
+            piety: 306,
+            skillspeed: 306,
+            spellspeed: 306,
+            tenacity: 306,
+
+            // Other
+            wdMag: 132,
+            wdPhys: 132,
+            weaponDelay: 0, // TODO: ?
+            vitality: 412,
+            hp: 0
+        });
+        eq(codexOfAscension.materiaSlots.length, 2);
+        eq(codexOfAscension.isCustomRelic, false);
+        // Not synced down - "Unsynced version" should just be the same
+        eq(codexOfAscension.unsyncedVersion, codexOfAscension);
+        eq(codexOfAscension.isUnique, true);
+        // eq(codexOfAscension.acquisitionType, 'raid');
+        eq(codexOfAscension.relicStatModel, undefined);
+
+        // This item should be filtered out due to being too low of an ilvl
+        // TODO: the new DataManager should just load everything, and filtering should be done visually
+        // const ilvl545book = dm.itemById(34691);
+        // eq(ilvl545book, undefined);
+
+        // This item is 560, it just barely makes it
+        const ilvl560book = dm.itemById(34053);
+        eq(ilvl560book.ilvl, 560);
+
+    }).timeout(20_000);
+    it('can get stats of food items', async () => {
+        const dm = new NewApiDataManager('SCH', 90);
         await dm.loadData();
         const food = dm.foodById(44096);
         eq(food.id, 44096);
