+++ packages/data-api-client/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@xivgear/data-api-client",
+  "version": "1.0.0",
+  "exports": {
+    "./*": "./src/*.ts"
+  },
+  "scripts": {
+    "build": "tsc --build"
+  },
+  "files": [
+    "package.json",
+    "tsconfig.json",
+    "src/*.ts"
+  ]
+}
