+++ packages/data-api-client/src/api_client.ts
@@ -1,0 +1,5 @@
+import {DataApiClient} from "./dataapi";
+
+new DataApiClient({
+    baseUrl: 'http://localhost:8085',
+});
